report/generate_report.py
from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer, Image, Table, TableStyle
from reportlab.lib.pagesizes import A4
from reportlab.lib import colors
from reportlab.lib.styles import getSampleStyleSheet
import json
import os

# ==== Config ====
metrics_path = "../results/metrics.json"
gradcam_path = "../results/gradcam_overlay.jpg"
sample_path = "../results/output_frame_sample.jpg"
output_pdf = "../results/UAV_Detection_Report.pdf"

# ==== Load dữ liệu ====
with open(metrics_path, "r") as f:
    metrics = json.load(f)

styles = getSampleStyleSheet()
style_title = styles["Title"]
style_body = styles["BodyText"]

doc = SimpleDocTemplate(output_pdf, pagesize=A4)
story = []

# ==== Tiêu đề ====
story.append(Paragraph("UAV Object Detection Report", style_title))
story.append(Spacer(1, 20))

# ==== Thông tin model ====
info = f"""
<b>Model:</b> {metrics["model_name"]}<br/>
<b>Dataset:</b> {metrics["dataset"]}<br/>
<b>Epochs:</b> {metrics["num_epochs"]}<br/>
"""
story.append(Paragraph(info, style_body))
story.append(Spacer(1, 10))

# ==== Bảng kết quả ====
data = [
    ["Metric", "Value"],
    ["Best mAP", f"{metrics['best_map']*100:.2f}%"],
    ["Average FPS", f"{metrics['avg_fps']:.2f}"],
    ["Average Latency (ms/frame)", f"{metrics['avg_latency_ms']:.2f}"]
]
table = Table(data, colWidths=[180, 180])
table.setStyle(TableStyle([
    ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
    ("TEXTCOLOR", (0, 0), (-1, 0), colors.black),
    ("ALIGN", (0, 0), (-1, -1), "CENTER"),
    ("BOX", (0, 0), (-1, -1), 1, colors.black),
    ("GRID", (0, 0), (-1, -1), 0.5, colors.grey)
]))
story.append(table)
story.append(Spacer(1, 20))

# ==== Hình minh họa ====
if os.path.exists(sample_path):
    story.append(Paragraph("<b>Example Inference Frame:</b>", style_body))
    story.append(Image(sample_path, width=400, height=225))
    story.append(Spacer(1, 20))

if os.path.exists(gradcam_path):
    story.append(Paragraph("<b>Grad-CAM Visualization:</b>", style_body))
    story.append(Image(gradcam_path, width=400, height=225))
    story.append(Spacer(1, 20))

# ==== Kết luận ====
story.append(Paragraph(
    "Model FasterRCNN-MobileNetV3 đã được fine-tune trên VisDrone-VID dataset. "
    "Kết quả cho thấy tốc độ trung bình đạt khoảng 14–15 FPS trên GPU, "
    "với độ chính xác mAP ~65%. Model có thể được cải thiện thêm bằng cách "
    "tăng kích thước input hoặc sử dụng data augmentation nâng cao.",
    style_body
))

# ==== Xuất PDF ====
doc.build(story)
print(f"✅ Báo cáo đã được tạo: {output_pdf}")
