requirements.txt
torch
torchvision
opencv-python
tqdm
matplotlib
numpy
Pillow
